+++ site/config.js
@@ -1,0 +1,3 @@
+window.FULL_CITY_CONFIG = {
+  fieldSubmissionEndpoint: ""
+};